Improved PitchMixer

+++ RocaloidEngine/CVE3/Synthesizer/CSynth.h
@@ -24,7 +24,7 @@ AutoClass
 
 typedef struct
 {
-    int PSOLAFrameHopSize;
+    float PSOLAFrameHopSize;
     int BeforeVOT;
 } CSynthSendback;
